Selenium_Python/requirements.txt
pytest == 7.1.2
selenium==4.2.0
allure-pytest==2.9.45